Update gamelogic.js
Commenting gameLogic to clear up confusions. Also added second ='s on line 55 to make while(is now true) become while(if true) 

Alternate implementation:
turnOrder var tracks whose turn it is, pushing a button when not your turn does nothing
playTurns() - takes an action (stand or hit), and a player (user or AI), then upon completion checks if next player is AI, and if so, determines their action based on an AI method, and then calls ITSELF with that player and the action.